suiteId拼写错误，生成excel。

+++ src/main/java/com/evtape/schedule/web/DistrictController.java
@@ -1,5 +1,7 @@
 package com.evtape.schedule.web;
 
+import java.util.List;
+
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
@@ -9,6 +11,7 @@ import org.springframework.web.bind.annotation.ResponseBody;
 import com.evtape.schedule.consts.ResultCode;
 import com.evtape.schedule.consts.ResultMap;
 import com.evtape.schedule.domain.District;
+import com.evtape.schedule.domain.Station;
 import com.evtape.schedule.persistent.Repositories;
 
 /**
@@ -72,7 +75,16 @@ public class DistrictController {
 		ResultMap resultMap;
 		try {
 			resultMap = new ResultMap(ResultCode.SUCCESS);
-			Repositories.districtRepository.delete(id);
+			List<Station> list = Repositories.stationRepository.findByDistrictId(id);
+			
+			if (list.size() > 0) {
+				resultMap.setData("当前站区下仍包含有站，不能删除此站区");
+			} else if (Repositories.districtRepository.exists(id)) {
+				Repositories.districtRepository.delete(id);
+				resultMap.setData("删除成功");
+			} else {
+				resultMap.setData("数据不存在，不用删除");
+			}
 		} catch (Exception e) {
 			resultMap = new ResultMap(ResultCode.SERVER_ERROR);
 		}
